+++ src/common/install/no_vnc.sh
@@ -5,9 +5,9 @@ set -u
 
 echo "Install noVNC - HTML5 based VNC viewer"
 mkdir -p $NO_VNC_HOME/utils/websockify
-wget -qO- https://github.com/novnc/noVNC/archive/v1.0.0.tar.gz | tar xz --strip 1 -C $NO_VNC_HOME
+wget -qO- https://github.com/novnc/noVNC/archive/refs/tags/v1.3.0.tar.gz | tar xz --strip 1 -C $NO_VNC_HOME
 # use older version of websockify to prevent hanging connections on offline containers, see https://github.com/ConSol/docker-headless-vnc-container/issues/50
-wget -qO- https://github.com/novnc/websockify/archive/v0.6.1.tar.gz | tar xz --strip 1 -C $NO_VNC_HOME/utils/websockify
-chmod +x -v $NO_VNC_HOME/utils/*.sh
+wget -qO- https://github.com/novnc/websockify/archive/refs/tags/v0.10.0.tar.gz | tar xz --strip 1 -C $NO_VNC_HOME/utils/websockify
+#chmod +x -v $NO_VNC_HOME/utils/*.sh
 ## create index.html to forward automatically to `vnc_lite.html`
 ln -s $NO_VNC_HOME/vnc_lite.html $NO_VNC_HOME/index.html