+++ src/debian/install/install_custom_fonts.sh
@@ -4,3 +4,4 @@ set -e
 
 echo "Installing ttf-wqy-zenhei"
 apt-get install -y ttf-wqy-zenhei
+apt-get clean -y