+++ src/debian/install/libnss_wrapper.sh
@@ -3,7 +3,7 @@
 set -e
 
 echo "Install nss-wrapper to be able to execute image as non-root user"
-apt-get update 
+apt-get update
 apt-get install -y libnss-wrapper gettext
 apt-get clean -y