+++ src/debian/install/tigervnc.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -e
+
+echo "Install TigerVNC server"
+apt-get install -y tigervnc-standalone-server
+apt-get clean -y