+++ src/rocky/install/tigervnc.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+set -e
+
+echo "Install TigerVNC server"
+dnf install -y tigervnc-server
+dnf clean all